analysis

+++ xip.c
@@ -585,7 +585,7 @@ ssize_t pmfs_xip_file_write(struct file *filp, const char __user *buf,
 		size_t hashing = 0;
 		void const *data_block = buf+count-i;
 		struct hash_map_addr *hash_map_addr_temp;
-		unsigned long *temp = kmalloc(sizeof(unsigned long), GFP_KERNEL);
+		unsigned long *temp = kmalloc(sizeof(size_t), GFP_KERNEL);
 		// char *xmem = kmalloc(pmfs_inode_blk_size(pi),GFP_KERNEL);
 		hash_map_addr_temp = kmalloc(sizeof(*hash_map_addr_temp), GFP_KERNEL);
